Fix 13 bugs found during code review
- EVERY/ENDEVERY: add parse-time fixup with end offset (like LOOP/TIMES),
  replacing broken linear scan that failed on nested blocks
- Parser: validate fixup kind in ELSE handler, distinguish DEF vs TASK
  blocks, replace unsafe mem::zeroed with safe init, make ParseResult
  const-generic
- Entry point: use Option<usize> instead of 0 sentinel
- I2C timeout: rename constant to clarify FreeRTOS tick units
- delay_ms: cap at 5 seconds to prevent unbounded blocking
- Scheduler: wire into run_program so TASK/START/STOP/ON/EMIT work
- bme_read: implement in Esp32Platform via I2C burst read
- wifi_ip: return packed i32 instead of string pool index
- task.rs: make Dict generic, propagate errors from process_actions
  and event dispatch, reset vars/every_last on task restart

+++ spore-core/src/platform.rs
@@ -75,7 +75,8 @@ pub trait Platform {
     fn wifi_disconnect(&mut self) -> PlatformResult<()> {
         Err(VmError::PlatformError)
     }
-    fn wifi_ip(&mut self) -> PlatformResult<u16> {
+    /// Returns the IP address as a packed i32 (network byte order).
+    fn wifi_ip(&mut self) -> PlatformResult<i32> {
         Err(VmError::PlatformError)
     }
